requirements.txt
# Pin diffusers to match vendored scripts/train_dreambooth_lora.py (see scripts/VENDORED.md).
# Install PyTorch separately for your CUDA version: https://pytorch.org/get-started/locally/
diffusers==0.31.0
accelerate>=0.33.0
# Pin below 5.x: newer transformers drop symbols (e.g. FLAX_WEIGHTS_NAME) that diffusers 0.31 imports.
transformers>=4.41.0,<5
peft>=0.11.0
huggingface_hub>=0.23.0
safetensors>=0.4.0
Pillow>=10.0.0
pillow-heif>=0.16.0
numpy>=1.24.0
packaging>=23.0
tqdm>=4.66.0
torchvision>=0.16.0
